% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abun.heatmap.g.R
\name{abun.heatmap.g}
\alias{abun.heatmap.g}
\title{Heat maps and statistical tests are drawn by group}
\usage{
abun.heatmap.g(ARG_type, group, Group_numb)
}
\arguments{
\item{ARG_type}{C.abundance, N.abundance, P.abundance, S.abundance}

\item{group}{Sample grouping file}

\item{Group_numb}{group number
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
result <- abun.heatmap.g(C.abundance,group,Group_numb)}
}
\description{
Heat maps and statistical tests are drawn by group
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com}
}
