% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abun.heatmap.s.R
\name{abun.heatmap.s}
\alias{abun.heatmap.s}
\title{Heat maps and statistical tests are drawn by sample}
\usage{
abun.heatmap.s(ARG_type, group, Sample_numb)
}
\arguments{
\item{ARG_type}{C.abundance, N.abundance, P.abundance, S.abundance}

\item{group}{Sample grouping file}

\item{Sample_numb}{sample number
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
result <- abun.heatmap.s(C.abundance,group,Sample_numb)}
}
\description{
Heat maps and statistical tests are drawn by sample
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com}
}
