% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anammox.NH4.N2H4.R
\name{Anammox.NH4.N2H4}
\alias{Anammox.NH4.N2H4}
\title{Anammox.NH4.N2H4 abundance information was extracted and combined at different species taxonomic levels}
\usage{
Anammox.NH4.N2H4(
  kegg.category = Gene,
  NR.tax = tax,
  profile = abundance,
  group = group
)
}
\arguments{
\item{kegg.category}{kegg grouping file}

\item{NR.tax}{A result of species annotation of gene sets using NR database}

\item{profile}{Results of abundance of gene sets}

\item{group}{Sample grouping file}
}
\description{
Anammox.NH4.N2H4 abundance information was extracted and combined at different species taxonomic levels
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
Anammox.NH4.N2H4 <- Anammox.NH4.N2H4(kegg.category = Gene,NR.tax = tax,profile = abundance,group = group)
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com}
}
