% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANR.NO3.NO2.R
\name{ANR.NO3.NO2}
\alias{ANR.NO3.NO2}
\title{ANR.NO3.NO2 abundance information was extracted and combined at different species taxonomic levels}
\usage{
ANR.NO3.NO2(
  kegg.category = Gene,
  NR.tax = tax,
  profile = abundance,
  group = group
)
}
\arguments{
\item{kegg.category}{kegg grouping file}

\item{NR.tax}{A result of species annotation of gene sets using NR database}

\item{profile}{Results of abundance of gene sets}

\item{group}{Sample grouping file}
}
\description{
ANR.NO3.NO2 abundance information was extracted and combined at different species taxonomic levels
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
ANR.NO3.NO2 <- ANR.NO3.NO2(kegg.category = Gene,NR.tax = tax,profile = abundance,group = group)
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com}
}
