% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARS.SO4.APS.R
\name{ARS.SO4.APS}
\alias{ARS.SO4.APS}
\title{ARS.SO4.APS abundance information was extracted and combined at different species taxonomic levels}
\usage{
ARS.SO4.APS(
  kegg.category = Gene,
  NR.tax = tax,
  profile = abundance,
  group = group
)
}
\arguments{
\item{kegg.category}{kegg grouping file}

\item{NR.tax}{A result of species annotation of gene sets using NR database}

\item{profile}{Results of abundance of gene sets}

\item{group}{Sample grouping file}
}
\description{
ARS.SO4.APS abundance information was extracted and combined at different species taxonomic levels
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)

ARS.SO4.APS <- ARS.SO4.APS(kegg.category = Gene,NR.tax = tax,profile = abundance,group = group)

}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com} Liang Zeng \email{zengliang@biozeron.com}
}
