% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocess.R
\name{Data.pre}
\alias{Data.pre}
\title{Data Preprocessing for kegg.profile.entry, group, kegg.category, NR.taxonomy, profile}
\usage{
Data.pre(
  kegg.entry = ko,
  group = group,
  kegg.category = Gene,
  NR.tax = tax,
  profile = abundance
)
}
\arguments{
\item{kegg.entry}{A profile table containing kegg's entry}

\item{group}{Sample grouping file}

\item{kegg.category}{kegg grouping file}

\item{NR.tax}{A result of species annotation of gene sets using NR database}

\item{profile}{Results of abundance of gene sets}
}
\description{
Data Preprocessing for kegg.profile.entry, group, kegg.category, NR.taxonomy, profile
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)

Data.pre(kegg.entry = ko,group = group,kegg.category = Gene,NR.tax = tax,profile = abundance)

}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com} Liang Zeng \email{zengliang@biozeron.com}
}
