% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Denitrification.N2O.N2.R
\name{Denitrification.N2O.N2}
\alias{Denitrification.N2O.N2}
\title{Denitrification.N2O.N2 abundance information was extracted and combined at different species taxonomic levels}
\usage{
Denitrification.N2O.N2(
  kegg.category = Gene,
  NR.tax = tax,
  profile = abundance,
  group = group
)
}
\arguments{
\item{kegg.category}{kegg grouping file}

\item{NR.tax}{A result of species annotation of gene sets using NR database}

\item{profile}{Results of abundance of gene sets}

\item{group}{Sample grouping file}
}
\description{
Denitrification.N2O.N2 abundance information was extracted and combined at different species taxonomic levels
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
Denitrification.N2O.N2 <- Denitrification.N2O.N2(kegg.category = Gene,NR.tax = tax,profile = abundance,group = group)
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com}
}
