% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Denitrification.NO.N2O.R
\name{Denitrification.NO.N2O}
\alias{Denitrification.NO.N2O}
\title{Denitrification.NO.N2O abundance information was extracted and combined at different species taxonomic levels}
\usage{
Denitrification.NO.N2O(
  kegg.category = Gene,
  NR.tax = tax,
  profile = abundance,
  group = group
)
}
\arguments{
\item{kegg.category}{kegg grouping file}

\item{NR.tax}{A result of species annotation of gene sets using NR database}

\item{profile}{Results of abundance of gene sets}

\item{group}{Sample grouping file}
}
\description{
Denitrification.NO.N2O abundance information was extracted and combined at different species taxonomic levels
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)

Denitrification.NO.N2O <- Denitrification.NO.N2O(kegg.category = Gene,NR.tax = tax,profile = abundance,group = group)

}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com} Liang Zeng \email{zengliang@biozeron.com}
}
