% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNR.NO2.NH4.R
\name{DNR.NO2.NH4}
\alias{DNR.NO2.NH4}
\title{DNR.NO2.NH4 abundance information was extracted and combined at different species taxonomic levels}
\usage{
DNR.NO2.NH4(
  kegg.category = Gene,
  NR.tax = tax,
  profile = abundance,
  group = group
)
}
\arguments{
\item{kegg.category}{kegg grouping file}

\item{NR.tax}{A result of species annotation of gene sets using NR database}

\item{profile}{Results of abundance of gene sets}

\item{group}{Sample grouping file}
}
\description{
DNR.NO2.NH4 abundance information was extracted and combined at different species taxonomic levels
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
DNR.NO2.NH4 <- DNR.NO2.NH4(kegg.category = Gene,NR.tax = tax,profile = abundance,group = group)
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com}
}
