% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host.ratio.R
\name{host.ratio}
\alias{host.ratio}
\title{Draw a heatmap and allow user to customize the title}
\usage{
host.ratio(host, title)
}
\arguments{
\item{host}{Host abundance of functional gene category}

\item{title}{Custom title for heat map}
}
\description{
Draw a heatmap and allow user to customize the title
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)

result <- host.ratio(AnCF[[1]],title)

}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com} Liang Zeng \email{zengliang@biozeron.com}
}
