% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ncycleabundance.R
\name{Ncyc.abundance}
\alias{Ncyc.abundance}
\title{ko abundance table related to nitrogen element was extracted}
\usage{
Ncyc.abundance(kegg.entry = ko)
}
\arguments{
\item{kegg.entry}{A profile table containing kegg's entry}
}
\description{
ko abundance table related to nitrogen element was extracted
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)

N.abundance = Ncyc.abundance(kegg.entry = ko)

}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com} Liang Zeng \email{zengliang@biozeron.com}
}
