% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ncycle.Hostrelative.stat.R
\name{Ncyc.host}
\alias{Ncyc.host}
\title{Determine whether a set of genes belongs to a specific biogeochemical cycle
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
result <- Ncyc.host(Gene)}
\usage{
Ncyc.host(Gene)
}
\description{
Determine whether a set of genes belongs to a specific biogeochemical cycle
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
result <- Ncyc.host(Gene)
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com}
}
