% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMDS.R
\name{nmds.arg}
\alias{nmds.arg}
\title{NMDS analysis and visualization}
\usage{
nmds.arg(ARG_sub, group)
}
\arguments{
\item{ARG_sub}{C.abundance[,2:ncol(C.abundance)], N.abundance[,2:ncol(N.abundance)], P.abundance[,2:ncol(P.abundance)], S.abundance[,2:ncol(S.abundance)]}

\item{group}{Sample grouping file
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
result <- nmds.arg(C.abundance[,2:ncol(C.abundance)],group)}
}
\description{
NMDS analysis and visualization
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com} Liang Zeng \email{zengliang@biozeron.com}
}
