% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NU.NH4.R
\name{NU.NH4}
\alias{NU.NH4}
\title{NU.NH4 abundance information was extracted and combined at different species taxonomic levels}
\usage{
NU.NH4(kegg.category = Gene, NR.tax = tax, profile = abundance, group = group)
}
\arguments{
\item{kegg.category}{kegg grouping file}

\item{NR.tax}{A result of species annotation of gene sets using NR database}

\item{profile}{Results of abundance of gene sets}

\item{group}{Sample grouping file}
}
\description{
NU.NH4 abundance information was extracted and combined at different species taxonomic levels
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
NU.NH4 <- NU.NH4(kegg.category = Gene,NR.tax = tax,profile = abundance,group = group)
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com}
}
