% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca.arg}
\alias{pca.arg}
\title{PCA analysis and visualization}
\usage{
pca.arg(ARG_sub, group)
}
\arguments{
\item{ARG_sub}{C.abundance[,2:ncol(C.abundance)], N.abundance[,2:ncol(N.abundance)], P.abundance[,2:ncol(P.abundance)], S.abundance[,2:ncol(S.abundance)]}

\item{group}{Sample grouping file
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
result <- pca.arg(C.abundance[,2:ncol(C.abundance)],group)}
}
\description{
PCA analysis and visualization
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com}
}
