% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pcycleabundance.R
\name{Pcyc.abundance}
\alias{Pcyc.abundance}
\title{ko abundance table related to phosphorus element was extracted}
\usage{
Pcyc.abundance(kegg.entry = ko)
}
\arguments{
\item{kegg.entry}{A profile table containing kegg's entry}
}
\description{
ko abundance table related to phosphorus element was extracted
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
P.abundance = Pcyc.abundance(kegg.entry = ko)
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com}
}
