% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scycleabundance.R
\name{Scyc.abundance}
\alias{Scyc.abundance}
\title{ko abundance table related to sulfur element was extracted}
\usage{
Scyc.abundance(kegg.entry = ko)
}
\arguments{
\item{kegg.entry}{A profile table containing kegg's entry}
}
\description{
ko abundance table related to sulfur element was extracted
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
S.abundance = Scyc.abundance(kegg.entry = ko)
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com} Liang Zeng \email{zengliang@biozeron.com}
}
