% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scycle.Hostrelative.stat.R
\name{Scyc.host}
\alias{Scyc.host}
\title{Determine whether a set of genes belongs to a specific biogeochemical cycle
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
result <- Scyc.host(Gene)}
\usage{
Scyc.host(Gene)
}
\description{
Determine whether a set of genes belongs to a specific biogeochemical cycle
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)
result <- Scyc.host(Gene)
}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com} Liang Zeng \email{zengliang@biozeron.com}
}
