% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ccycle.Hostrelative.stat.R
\name{Ccyc.host}
\alias{Ccyc.host}
\title{Determine whether a set of genes belongs to a specific biogeochemical cycle}
\usage{
Ccyc.host(Gene)
}
\description{
Determine whether a set of genes belongs to a specific biogeochemical cycle
}
\examples{
data(ko)
data(group)
data(Gene)
data(tax)
data(abundance)

result <- Ccyc.host(Gene)

}
\author{
contact: Zhengfu Yue \email{yuezhengfu2011@163.com} Liang Zeng \email{zengliang@biozeron.com}
}
